rag-backend/requirements.txt
fastapi
uvicorn
python-docx
pdfplumber
python-dotenv
qdrant-client
sentence-transformers